Refactoring bulk update on establishments local identifiers, and optimising the updates to only those that have changed and not to updatte the establishment's auditing.
