+++ server/routes/establishments/localIdentifier.js
@@ -81,59 +81,6 @@ router.route('/').post(async (req, res) => {
   }
 });
 
-
-// a helper function that updates the establishment and adds the necessary audit events
-const updateLocalIdOnEstablishment = async (thisGivenEstablishment, transaction, updatedTimestamp, username, allAuditEvents) => {
-
-  const updatedEstablishment = await models.establishment.update(
-    {
-      LocalIdentifierValue: thisGivenEstablishment.value,
-      LocalIdentifierSavedBy: username,
-      LocalIdentifierChangedBy: username,
-      LocalIdentifierSavedAt: updatedTimestamp,
-      LocalIdentifierChangedAt: updatedTimestamp,
-      updated: updatedTimestamp,
-      updatedBy: username,
-    },
-    {
-      returning: true,
-      where: {
-          uid: thisGivenEstablishment.uid
-      },
-      attributes: ['id', 'updated'],
-      transaction,
-    }
-  );
-
-  if (updatedEstablishment[0] === 1) {
-    const updatedRecord = updatedEstablishment[1][0].get({plain: true});
-    // and now the audit events - one for the establishment entity
-    allAuditEvents.push({
-      establishmentFk: updatedRecord.EstablishmentID,
-      username,
-      type: 'updated'
-    });
-
-    // and two for the LocalIdentifier property (saved and changed)
-    allAuditEvents.push({
-      establishmentFk: updatedRecord.EstablishmentID,
-      username,
-      type: 'saved',
-      property: 'LocalIdentifier',
-    });
-    allAuditEvents.push({
-      establishmentFk: updatedRecord.EstablishmentID,
-      username,
-      type: 'changed',
-      property: 'LocalIdentifier',
-      event: {
-        new: thisGivenEstablishment.value
-      }
-    });
-  }
-
-}
-
 // a workaround to allow updating all local identifiers for all given establishments in one transaction
 router.route('/').put(async (req, res) => {
   const establishmentId = req.establishmentId;
@@ -151,43 +98,9 @@ router.route('/').put(async (req, res) => {
   try {
     // as a minimum for security purposes, we restore the user's primary establishment
     if (await thisEstablishment.restore(establishmentId)) {
+      const updatedUids = await thisEstablishment.bulkUpdateLocalIdentifier(username, givenLocalIdentifiers);
 
-      // having restored their primary establishment and hence authenticated, we also fetch a list the user's current set of establishments
-      const thisUser = new User(establishmentId);
-      await thisUser.restore(null, username,false);
-      const myEstablishments = await thisUser.myEstablishments(isParent);
-
-      // create a list of those establishment UIDs - the user will only be able to update the local identifier for which they own
-      const myEstablishmentUIDs = [];
-      myEstablishmentUIDs.push(myEstablishments.primary.uid);
-
-      if (myEstablishments.subsidaries) {
-        myEstablishments.subsidaries.establishments.forEach(thisEst => {
-          myEstablishmentUIDs.push(thisEst.uid);
-        });
-      }
-
-      // within one transaction
       const updatedTimestamp = new Date();
-      const updatedUids = [];
-      await models.sequelize.transaction(async t => {
-        const dbUpdatePromises = [];
-        const allAuditEvents = [];
-        givenLocalIdentifiers.forEach(thisGivenEstablishment => {
-          if (thisGivenEstablishment && thisGivenEstablishment.uid && myEstablishmentUIDs.includes(thisGivenEstablishment.uid)) {
-            const updateThisEstablishment = updateLocalIdOnEstablishment(thisGivenEstablishment, t, updatedTimestamp, username, allAuditEvents);
-            dbUpdatePromises.push(updateThisEstablishment);
-            updatedUids.push(thisGivenEstablishment);
-          } else {
-            // simply ignore it - silient operation, as the frontend is not going to give dodgy UIDs
-          }
-        });
-
-        // wait for all updates to finish
-        await Promise.all(dbUpdatePromises);
-        await models.establishmentAudit.bulkCreate(allAuditEvents, {transaction: t});
-      });
-
       return res.status(200).json({
         id: thisEstablishment.id,
         uid: thisEstablishment.uid,
@@ -210,6 +123,5 @@ router.route('/').put(async (req, res) => {
     return res.status(503).send(err.message);
   }
 });
-
 
 module.exports = router;
